Streamlit/upload_file_to_db/requirements.txt
streamlit
pandas 
pillow 
psycopg2-binary==2.9.10
SQLAlchemy

